Move nlp tools back into single file
