+++ adeft/score/_score.pyx
@@ -7,8 +7,6 @@ cdef extern from 'limits.h':
 from cython cimport boundscheck, wraparound
 from cpython.mem cimport PyMem_Malloc, PyMem_Free
 
-from adeft.nlp.resources import stopwords_min
-
 include 'permutations.pyx'
 
 logger = logging.getLogger(__name__)
